chore (repo adjustment): Adjusted NextJS project to follow default/recommended settings + Prisma ORM
A major change was released before we start developing our components.
This change ensures we are follwoing NextJS recommendations and also
introduces basic DB configuration (readme file has been updated to
include examples on how to use it where to find more tutorials).
Note: we are still using MongoDB, just using a different way to
manipulate it.

With this change we can be confident developing the UI. I have not
included Tailwind libraries yet, but that will come after I deploy the
landing page.

From this point I will also be including a dev branch so we can collab.
More info on Readme.md.